refactor: rename module modifications.interfaces into modifications.specification to better reflect it current purpose
