+++ src/dfd/datasets/modifications/definitions/gaussian_noise.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from dfd.datasets.modifications.interfaces import ModificationSpecification
+from dfd.datasets.modifications.specification import ModificationSpecification
 
 
 class GaussianNoiseModification(ModificationSpecification):
