+++ src/dfd/datasets/modifications/definitions/histogram_equalization.py
@@ -3,7 +3,7 @@
 import cv2 as cv
 import numpy as np
 
-from dfd.datasets.modifications.interfaces import ModificationSpecification
+from dfd.datasets.modifications.specification import ModificationSpecification
 
 
 class HistogramEqualizationModification(ModificationSpecification):
