+++ src/dfd/datasets/modifications/definitions/identity.py
@@ -1,7 +1,7 @@
 """Modification Identity."""
 import numpy as np
 
-from dfd.datasets.modifications.interfaces import ModificationSpecification
+from dfd.datasets.modifications.specification import ModificationSpecification
 
 
 class IdentityModification(ModificationSpecification):
